+++ cmd/get.go
@@ -61,7 +61,7 @@ func getAudioByID(cmd *cobra.Command, verbose bool) ([]byte, error) {
 		}
 	}
 	params := "id=" + url.QueryEscape(id)
-	path := fmt.Sprintf("http://%s:%d/request?%s", viper.Get("cli.hostname"), int(viper.Get("cli.port").(float64)), params)
+	path := fmt.Sprintf("http://%s:%d/request?%s", viper.Get("cli.hostname"), viper.Get("cli.port").(int), params)
 	payload := &bytes.Buffer{}
 	req, err := http.NewRequest(http.MethodGet, path, payload)
 	if err != nil {
